requirements.txt
blinker==1.6.3
boto3==1.16.63
botocore==1.19.63
bson==0.5.10
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.3
python-dateutil==2.8.1
python-dotenv==0.15.0
s3transfer==0.3.4
six==1.15.0
urllib3==1.26.3
Werkzeug==3.0.0
